Separate server impl from grpc service impl
